+++ src/main/java/br/com/fatecmogidascruzes/SistemaDeLivraria.java
@@ -5,29 +5,27 @@ import java.time.LocalDate;
 import br.com.fatecmogidascruzes.model.entity.Livro;
 import br.com.fatecmogidascruzes.model.repository.LivroRepository;
 import br.com.fatecmogidascruzes.service.impl.LivroServiceImpl;
+import br.com.fatecmogidascruzes.validator.LivroValidator;
 
 public class SistemaDeLivraria {
     public static void main(String[] args) {
-    	LivroRepository livroRepository = new LivroRepository();
-    	LivroServiceImpl service = new LivroServiceImpl(livroRepository);
-        LocalDate dataDePublicacao = LocalDate.of(0, 1, 1); // Ano 0000, Janeiro 1
-    	
+
+        LocalDate dataDePublicacao = LocalDate.of(0, 1, 1);
+
+        LivroRepository livroRepository = new LivroRepository();
+        LivroValidator livroValidator = new LivroValidator();
+        LivroServiceImpl service = new LivroServiceImpl(livroRepository, livroValidator);
+
         Livro livro = new Livro(1, "0123456789", "", "Biblia", "Portugues", "Jesus", 7, "Reino do Ceus", 1000, dataDePublicacao, 300.00, "Religiao");
-    	
+
         service.adicionarLivro(livro);
         
-        System.out.println("ID: " + livro.getId());
-        System.out.println("ISBN-10: " + livro.getIsbn10());
-        System.out.println("ISBN-13: " + livro.getIsbn13());
-        System.out.println("Título: " + livro.getTitulo());
-        System.out.println("Idioma: " + livro.getIdioma());
-        System.out.println("Autores: " + livro.getAutores());
-        System.out.println("Estoque: " + livro.getEstoque());
-        System.out.println("Editora: " + livro.getEditora());
-        System.out.println("Número de Páginas: " + livro.getNumeroDePaginas());
-        System.out.println("Data de Publicação: " + livro.getDataDePublicacao());
-        System.out.println("Preço: " + livro.getPreco());
-        System.out.println("Categoria: " + livro.getCategoria());
-    	
+        System.out.println(" ID: " + livro.getId() + "\n Título: " + livro.getTitulo() + "\n Autores: " + livro.getAutor() + "\n Estoque: " + livro.getEstoque() + "\n Categoria: " + livro.getCategoria());
+
+        livro = new Livro(1, "0123456789", "", "Biblia", "Portugues", "profeta", -1, "Reino do Ceus", 1000, dataDePublicacao, 300.00, "Religiao");
+
+        service.alterarLivro(livro);
+
+        System.out.println(" ID: " + livro.getId() + "\n Título: " + livro.getTitulo() + "\n Autores: " + livro.getAutor() + "\n Estoque: " + livro.getEstoque() + "\n Categoria: " + livro.getCategoria());
     }
 }
