+++ src/main/java/br/com/fatecmogidascruzes/model/entity/Cliente.java
@@ -1,14 +1,5 @@
 package br.com.fatecmogidascruzes.model.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
 public class Cliente extends Usuario {
 
 }
