+++ src/main/java/br/com/fatecmogidascruzes/service/LivroService.java
@@ -5,6 +5,5 @@ import br.com.fatecmogidascruzes.model.entity.Livro;
 public interface LivroService {
 	
 	void adicionarLivro(Livro livro);
-	boolean validarAdicionar(Livro livro);
-	
+	void alterarLivro(Livro livro);
 }
